LocalWalFile.Load now return uint32 rather than custom type

+++ pkg/service/db.go
@@ -61,7 +61,7 @@ func (r *DBListRepo) Start(client Client, walChan chan *[]EventLog, inputEvtsCha
 
 // Stop is called on app shutdown. It flushes all state changes in memory to disk
 func (r *DBListRepo) Stop() error {
-	err := r.LocalWalFile.Stop(r.uuid)
+	err := r.LocalWalFile.Stop(uint32(r.uuid))
 	if err != nil {
 		return err
 	}
